Add files via upload
*****WELCOME TO MY PROGRAMMING README PART**************************************
This game basically plays between the player and the computer. It is a game between the player 
and the computer, please check the results as they flash on the top part.
For more of artificial intelligence codes and projects, you can share or 
follow my github page to get more. Hope you enjoy playing this game. 


****************************NICE PLAYING******************************************